% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calba.R
\name{ba_decay}
\alias{ba_decay}
\title{Calculate Basal Area with Decay Function}
\usage{
ba_decay(mu_values, data, n_focal, r)
}
\arguments{
\item{mu_values}{A numeric vector of decay parameters. Each value in \code{mu_values} represents
a decay factor used in calculating basal area for different focal trees.}

\item{data}{A data frame containing the tree data, including columns such as \code{gx}, \code{gy}, and \code{ba}
for the x-coordinates, y-coordinates, and basal area of each tree.}

\item{n_focal}{An integer specifying the number of focal trees to include in the calculation.}

\item{r}{A numeric scalar representing a decay radius or scaling factor, used to modify
the decay effect in the calculations.}
}
\value{
A list with two matrices:
\describe{
\item{\code{con_ba_matrix}}{A numeric matrix of basal areas with decay applied for conspecific (same species) trees.}
\item{\code{het_ba_matrix}}{A numeric matrix of basal areas with decay applied for heterospecific (different species) trees.}
}
}
\description{
This function calculates the basal area across a given set of trees, adjusting for decay based
on mu values (decay parameters). It uses a zero-truncated compound Poisson-lognormal distribution
mixture model to handle multiple focal trees with specified decay.
}
\details{
This function applies a decay model to adjust basal area values based on distance
and species type. The compound Poisson-lognormal distribution mixture is used to handle
zero-truncation, ensuring no zero basal area values.

This is useful for ecological studies where the effect of tree proximity (distance-decay) and
tree type (same vs. different species) are important for understanding basal area distribution.
}
\examples{
# Generate a sample data frame
sample_data <- data.frame(latin = sample(letters[1:4], 100, replace = TRUE), gx = runif(100), gy = runif(100), ba = runif(100, 10, 30))
mu_values <- c(1, 3, 5, 7)
ba_decay(mu_values, sample_data, n_focal = 10, r = 1.5)

}
