% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calba.R
\name{ba_simple}
\alias{ba_simple}
\title{Calculate Simple Basal Area}
\usage{
ba_simple(dbh, n_focal, r)
}
\arguments{
\item{dbh}{A numeric vector of tree diameters at breast height (DBH) in centimeters.}

\item{n_focal}{An integer specifying the number of focal trees to include in the calculation.
This parameter limits the number of values from `dbh` that are used.}

\item{r}{A numeric scalar for the radius parameter. This parameter could represent a distance
or scaling factor used in calculating the basal area (interpretation depends on context).}
}
\value{
A numeric vector representing the basal area for each specified tree in square centimeters.
  The length of the output vector is equal to `n_focal`.
}
\description{
This function calculates the basal area for a given set of tree diameters at breast height (DBH),
the number of focal trees, and a radius parameter.
}
\details{
The basal area is a measure commonly used in forestry and ecology to represent the cross-sectional
area of a tree trunk at breast height, typically at 1.3 meters above the ground. The function
uses a simple formula to calculate basal area for multiple trees.

The formula used to calculate the basal area is based on the approximation:
  \deqn{\text{basal area} = \pi \cdot \left(\frac{\text{dbh}}{2}\right)^2}
  where DBH is the diameter at breast height.
}
\examples{
# Calculate basal area for a set of DBH values
sample_data <- data.frame(latin = sample(letters[1:4], 100, replace = TRUE), gx = runif(100), gy = runif(100), ba = runif(100, 10, 30))
ba_simple( sample_data, n_focal = 10, r = 1.5)

}
