% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calba.R
\name{count_con}
\alias{count_con}
\title{Count Conspecific Trees}
\usage{
count_con(sp, gx, gy, r)
}
\arguments{
\item{sp}{A character vector of species names.}

\item{gx}{A numeric vector of x-coordinates for the trees.}

\item{gy}{A numeric vector of y-coordinates for the trees.}

\item{r}{A numeric scalar for the radius parameter.}
}
\value{
A numeric vector containing the count of conspecific trees within the radius for each focal tree.
}
\description{
This function counts the number of conspecific trees within a given radius for each focal tree.
}
\examples{
sample_data <- data.frame(
  latin = sample(letters[1:4], 100, replace = TRUE),
  gx = runif(100),
  gy = runif(100)
)
count_con(
  sp = sample_data$latin,
  gx = sample_data$gx,
  gy = sample_data$gy,
  r = 1.5
)

}
